TBCC: extension gallery UX, capture heuristics, dashboard/backend updates
- Extension: OnlyFans capture/thumbnail fixes, overlay metadata, sub-tab history,
  grid sort/details view, lightbox video errors, flicker/thumb proxy tweaks
- Backend: link resolver, payment bot settings, watch folder API, migrations,
  secretary bot scaffold, subscription/payment/scheduling adjustments
- Dashboard: PaymentBotSettings, WatchFolder, media/scheduler/bot panels

+++ tbcc/backend/app/models/external_payment_order.py
@@ -1,6 +1,6 @@
 """Pending external (wallet / manual) payments before admin verification."""
 
-from sqlalchemy import Column, DateTime, ForeignKey, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, ForeignKey, Integer, String
 
 from .base import Base
 
@@ -9,7 +9,7 @@ class ExternalPaymentOrder(Base):
     __tablename__ = "external_payment_orders"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_user_id = Column(Integer, nullable=False, index=True)
+    telegram_user_id = Column(BigInteger, nullable=False, index=True)
     plan_id = Column(Integer, ForeignKey("subscription_plans.id"), nullable=False)
     reference_code = Column(String(32), unique=True, nullable=False, index=True)
     status = Column(String(16), nullable=False, default="pending")  # pending|paid|cancelled
