+++ tbcc/backend/app/models/referral_code.py
@@ -1,6 +1,6 @@
 """Persistent unique referral codes per Telegram user (short ref_* deep links)."""
 
-from sqlalchemy import Column, DateTime, Integer, String
+from sqlalchemy import BigInteger, Column, DateTime, String
 
 from .base import Base
 
@@ -8,6 +8,6 @@ from .base import Base
 class ReferralCode(Base):
     __tablename__ = "referral_codes"
 
-    telegram_user_id = Column(Integer, primary_key=True, autoincrement=False)
+    telegram_user_id = Column(BigInteger, primary_key=True, autoincrement=False)
     code = Column(String(16), unique=True, nullable=False, index=True)
     created_at = Column(DateTime, nullable=True)
