+++ tbcc/backend/app/models/subscription.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy import BigInteger, Column, Integer, String, DateTime, ForeignKey
 
 from .base import Base
 
@@ -7,13 +7,13 @@ class Subscription(Base):
     __tablename__ = "subscriptions"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    telegram_user_id = Column(Integer, nullable=False)
+    telegram_user_id = Column(BigInteger, nullable=False)
     plan_id = Column(Integer, ForeignKey("subscription_plans.id"), nullable=True)
     plan = Column(String, nullable=True)  # denormalized plan name for display
     status = Column(String, default="active")  # active|expired|cancelled
     expires_at = Column(DateTime, nullable=True)
     payment_method = Column(String, nullable=True)  # stars|crypto|manual
     amount_stars = Column(Integer, nullable=True)  # Stars paid at purchase (for analytics)
-    referrer_id = Column(Integer, nullable=True)  # Telegram user who referred this subscriber
+    referrer_id = Column(BigInteger, nullable=True)  # Telegram user who referred this subscriber
     # Idempotent Stars fulfillment — duplicate successful_payment webhooks return same row
     telegram_payment_charge_id = Column(String(128), nullable=True, unique=True)
